Fixed compression bug.
This bug occurred because the way the code worked, there could be
repeated runs of one byte, which isn't really allowed by the game. So we
split them up first, and let the existing code work with what is no
longer a repeat of length 1, but a single, solitary byte.