Tweak/bugfix to traffic compositing.
In the original game, traffic only draws when it's over something road
coloured, so that it goes under powerlines, whereas it originally went
over in this code. Now it goes under.
And in keeping with the original game, traffic also goes under the
traintracks and crosswalks.

I'm still not sure the thresholds for traffic changing are correct.