Highway rendering is better...
This required a bit of a refactor to how buildings are found, because
conceptually a 2x2 bighway tile is a building like other and need to be
handled differently than 1x1 network tiles.

This still only kinda works, there's some height offsets that need to be
fixed. I forsee a lot of work special casing the highway drawing, and it
feels like I'm missing something subtle here, but I can't quite figure
it out.