First pass at scenario creation
It actually mostly appears to work, but is a bit clunky to use.
The only outstanding issue that remains is the colours are wrong for the converted PICT, which looks like a similar indexing issue I had converting the other way. Should (hopefully) be an easy fix.

Added serialization of the scenario to sc2_parse as well, including hacky fixes for the duplicated TEXT segments.