PICT conversion working, cleanup
Conversion from an image to a PICT image works now, though the colour remapping function could use a little work, it doesn't seem to pick the best colours, so there may be a bug, or it may not be the best algorithm to use.

Also cleaned up the serialization function of the scenario, and also moved some of the code from sc2_parse to sc2_serialize, where it logically makes more sense to be. Potentially the whole city.serialize() function should actually live there...