PICT tweaks.
Redid serialization slightly.
Also changed the colour choosing logic to at least allow proper round-tripping of PICT files extracted from existing scenarios.
It's still not quite making images that look good for arbitrary images, but this may just be a limitation of the palette versus images made with that palette in mind.