Added first-pass sign rendering.
Currently trying font "W95FA.otf", available online. I think the font used in the Win95 version is probably MS Sans Serif, and this is relatively close, though it doesn't seem to support a bunch of the characters.
There are no bitmaps included for the signs, I hand coded them. This may change at some point, but this is really trying to emulate a long-gone UI toolkit, I think.