Change ALTM parsing.
Due to a PR on the sc2k-docs repo, I now understand what some unknown bits are for, the depth of the water. Added this to ALTM parsing.

This also allowed me to figure out what was going on with a years old bug in the game, where certain cities had an offset from the cursor position to the tile the game thought it was on. This is due to a bug in the game's terrain generation causing the water depth to get out of sync with the altitude when on dry land. I added a --fix-offset mode to city_tools.py to fix this, and it seems to work.